// src/generated/prisma/client/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  ADMIN: 'ADMIN',
  MANAGER: 'MANAGER',
  FARMER: 'FARMER'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const ItemCategory = {
  FEED: 'FEED',
  MEDICINE: 'MEDICINE',
  CHICKS: 'CHICKS',
  OTHER: 'OTHER'
} as const

export type ItemCategory = (typeof ItemCategory)[keyof typeof ItemCategory]


export const TransactionType = {
  PURCHASE: 'PURCHASE',
  PAYMENT: 'PAYMENT'
} as const

export type TransactionType = (typeof TransactionType)[keyof typeof TransactionType]


export const PaymentMethod = {
  CASH: 'CASH',
  BANK: 'BANK',
  MOBILE_BANKING: 'MOBILE_BANKING',
  OTHER: 'OTHER'
} as const

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod]


export const AuditAction = {
  CREATE: 'CREATE',
  UPDATE: 'UPDATE',
  DELETE: 'DELETE'
} as const

export type AuditAction = (typeof AuditAction)[keyof typeof AuditAction]